+++ R/data.R
@@ -89,6 +89,19 @@
 #' ggplot(lhd) + geom_sf(aes(fill = lhd_name), show.legend = FALSE)
 "lhd"
 
+#' Primary Health Network boundaries of New South Wales
+#'
+#' The geometries have been simplified with a tolerance of 500 m to reduce the
+#' level of detail.
+#'
+#' @source
+#'   National Recovery and Resilience Agency, "PHN Boundaries used by the NBRA",
+#'   \url{https://data.gov.au/data/dataset/phn-boundaries-used-by-the-nbra}, accessed 25 May 2023.
+#'
+#'   The original dataset is published under the [Creative Commons Attribution 2.5 Australia](https://creativecommons.org/licenses/by/2.5/au/) licence,
+#'   © Commonwealth of Australia 2021.
+"phn"
+
 #' Postal codes and localities of New South Wales.
 #'
 #' Derived from several government sources with some community curation.