+++ NEWS.md
@@ -1,5 +1,7 @@
 # nswgeo (development version)
 
+* Add Public Health Network (PHN) map data as `phn` dataset.
+
 # nswgeo 0.3.3
 
 * Initial CRAN release.